// src/generated/types/CreateEventDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export const createEventDtoEventTypeEnum = {
  exam: 'exam',
  quiz: 'quiz',
  assignment: 'assignment',
  lab_assignment: 'lab_assignment',
  project: 'project',
  presentation: 'presentation',
  workshop: 'workshop',
  practice: 'practice',
  seminar: 'seminar',
} as const

export type CreateEventDtoEventTypeEnum = (typeof createEventDtoEventTypeEnum)[keyof typeof createEventDtoEventTypeEnum]

export const createEventDtoLocationTypeEnum = {
  lab_devices: 'lab_devices',
  lecture_hall: 'lecture_hall',
  online: 'online',
  hybrid: 'hybrid',
} as const

export type CreateEventDtoLocationTypeEnum = (typeof createEventDtoLocationTypeEnum)[keyof typeof createEventDtoLocationTypeEnum]

export type CreateEventDto = {
  /**
   * @type string
   */
  name: string
  /**
   * @type string | undefined
   */
  description?: string | undefined
  /**
   * @type number
   */
  duration: number
  /**
   * @default "assignment"
   * @type string
   */
  eventType: CreateEventDtoEventTypeEnum
  /**
   * @default "online"
   * @type string
   */
  locationType: CreateEventDtoLocationTypeEnum
  /**
   * @description Custom location when not using lab devices
   * @type string | undefined
   */
  customLocation?: string | undefined
  /**
   * @type boolean
   */
  hasMarks: boolean
  /**
   * @type number | undefined
   */
  totalMarks?: number | undefined
  /**
   * @default false
   * @type boolean
   */
  autoStart: boolean
  /**
   * @default 30
   * @type number
   */
  examModeStartMinutes: number
  /**
   * @description When the event should start
   * @type string | undefined, date-time
   */
  startDateTime?: Date | undefined
  /**
   * @description Whether this event requires exam models
   * @default false
   * @type boolean
   */
  requiresModels: boolean
  /**
   * @description Allow random assignment of exam models to students
   * @default false
   * @type boolean
   */
  allowRandomModelAssignment: boolean
  /**
   * @description Whether this event should be treated as an exam
   * @default false
   * @type boolean
   */
  isExam: boolean
  /**
   * @type string
   */
  courseId: string
}